requirements.txt
torch>=2.4.0
diffusers>=0.37.0
transformers>=5.3.0
accelerate>=1.13.0
peft>=0.10.0
xformers>=0.0.24
bitsandbytes>=0.43.0
safetensors>=0.4.2
tensorboard>=2.16.0
opencv-python>=4.9.0
albumentations>=1.3.0
omegaconf>=2.3.0
Pillow>=10.0.0
tqdm>=4.66.0
scipy>=1.11.0
